Module Shuttle

module Input_channel : sig ... end

Alternative to Async_unix.Reader, based on the low latency transport in async_rpc.

module Output_channel : sig ... end

Alternative to Async_unix.Writer, based on the low latency transport in async_rpc.

module Connection : sig ... end