Module Output_channel.Flush_result

type t =

| Flushed

| Remote_closed

| Error

t Result of a flush operation.

Flushed indicates all prior writes at the time flush was call have finished without any errors.

Remote_closed indicates that the consumer that's reading the bytes written to the Output_channel is closed, i.e. the channel has received an EPIPE or ECONNRESET when it attempts to perform a write.

Error indicates that the write operation was interrupted by an unhandled exception, or a timeout.

val sexp_of_t : t -> Sexplib0.Sexp.t